fix: グローバルステート theme の宣言場所を変更
